+++ src/main/java/xyz/valnet/hadean/input/GOButton.java
@@ -57,7 +57,7 @@ public class GOButton extends GameObject implements IMouseCaptureArea {
   }
 
   @Override
-  public int getLayer() {
+  public float getLayer() {
     return layer;
   }
